+++ src/i18n/routing.ts
@@ -1,9 +1,10 @@
 import { defineRouting } from 'next-intl/routing'
 import { SECONDS_PER_YEAR } from '@/lib/datetime/constants'
+import { DEFAULT_LOCALE, LOCALES } from './locales'
 
 export const routing = defineRouting({
-  locales: ['en', 'de'],
-  defaultLocale: 'de',
+  locales: [...LOCALES],
+  defaultLocale: DEFAULT_LOCALE,
   localePrefix: 'never',
   localeCookie: {
     name: 'NEXT_LOCALE',